Finalise deque (#26)
* update Justfile

* group lists and queues

* finalise Deque

* add v0.5.0 release notes
