+++ python/adspy/data_structures/lists.py
