+++ python/tests/data_structures/test_deque.py
@@ -1,0 +1,328 @@
+import copy
+import random
+from collections import deque
+from collections.abc import Iterable, MutableSequence
+from contextlib import AbstractContextManager
+from contextlib import nullcontext as does_not_raise
+from sys import maxsize
+from typing import Any
+
+import pytest
+
+from adspy.data_structures import Deque
+
+
+def test_is_mutable_sequence() -> None:
+    dq = Deque()
+    for ms in (deque(), dq):  # type: ignore [var-annotated]
+        assert isinstance(ms, MutableSequence)
+        assert issubclass(type(ms), MutableSequence)
+
+
+@pytest.mark.parametrize(
+    ("it", "maxlen", "answer"),
+    [
+        ([], None, []),
+        ([], 0, []),
+        ([], 1, []),
+        ([42, 21], None, [42, 21]),
+        ([42, 21], 0, []),
+        ([42, 21], 1, [21]),
+        ([42, 21], 2, [42, 21]),
+    ],
+)
+def test_init_with_maxlen(it: Iterable, maxlen: int, answer: list) -> None:
+    assert Deque(it, maxlen=maxlen) == answer
+
+
+@pytest.mark.parametrize("elements", [[1], (2, 3), "456"])
+def test_append(elements: Iterable) -> None:
+    items = list(elements)
+    dq = Deque()
+
+    ans = items
+    for value in items:
+        dq.append(value)
+
+    assert dq == ans
+    assert len(dq) == len(ans)
+
+
+def test_append_with_maxlen_one() -> None:
+    dq = Deque(maxlen=1)
+
+    dq.append(1)
+    dq.append(21)
+
+    assert dq == [21]
+
+    dq.append(42)
+    assert dq == [42]
+
+
+@pytest.mark.parametrize("elements", [[1], (2, 3), "456"])
+def test_extend(elements: Iterable) -> None:
+    items = list(elements)
+    dq = Deque()
+
+    ans = items
+    dq.extend(ans)
+
+    assert dq == items
+    assert len(dq) == len(ans)
+
+
+def test_extend_with_maxlen_one() -> None:
+    items = [1, 21, 42]
+    dq = Deque(maxlen=1)
+
+    dq.extend(items)
+
+    assert dq == [42]
+    assert Deque(items, maxlen=1) == [42]
+
+
+@pytest.mark.parametrize("elements", [[1], (2, 3), "456"])
+def test_appendleft(elements: Iterable) -> None:
+    items = list(elements)
+    dq = Deque()
+
+    ans = items
+    for value in items:
+        dq.appendleft(value)
+
+    assert dq == list(reversed(ans))
+    assert len(dq) == len(ans)
+
+
+def test_appendleft_with_maxlen_one() -> None:
+    dq = Deque(maxlen=1)
+
+    dq.appendleft(1)
+    dq.appendleft(21)
+
+    assert dq == [21]
+
+    dq.appendleft(42)
+    assert dq == [42]
+
+
+@pytest.mark.parametrize("elements", [[1], (2, 3), "456"])
+def test_extendleft(elements: Iterable) -> None:
+    items = list(elements)
+    dq = Deque()
+
+    ans = items
+    dq.extendleft(ans)
+
+    assert dq == list(reversed(items))
+    assert len(dq) == len(ans)
+
+
+def test_extendleft_with_maxlen_one() -> None:
+    items = [1, 21, 42]
+    dq = Deque(maxlen=1)
+
+    dq.extendleft(items)
+
+    assert dq == [42]
+    assert Deque(items, maxlen=1) == [42]
+
+
+@pytest.mark.parametrize("it", [[], [1], [-1, 0, 1]])
+def test_clear(it: Iterable) -> None:
+    dq = Deque(it)
+
+    dq.clear()
+
+    assert not len(dq)
+    assert not tuple(dq)
+
+    with pytest.raises(IndexError):
+        dq[0]
+
+
+@pytest.mark.parametrize("lst", [[], [1, "2", [3, [4, 5]]]])
+def test_copy(lst: list) -> None:
+    dq = Deque(lst)
+    ans = deque(lst)
+
+    copied = copy.copy(dq)
+    assert isinstance(copied, Deque)
+    assert copied == ans
+
+    copied = dq.copy()
+    assert isinstance(copied, Deque)
+    assert copied == ans
+
+
+@pytest.mark.parametrize(
+    ("lst", "value"),
+    [
+        ([], 1),
+        ([1, 2, 1, 0, 1, 3], 1),
+    ],
+)
+def test_count(lst: list, value: Any) -> None:
+    dq = Deque(lst)
+    assert dq.count(value) == deque(lst).count(value)
+
+
+@pytest.mark.parametrize(
+    ("lst", "value", "start", "stop", "expectation"),
+    [
+        ([], 1, 0, maxsize, pytest.raises(ValueError)),  # noqa: PT011
+        ([2, 1, 3], 1, 1, maxsize, does_not_raise()),
+        ([2, 1, 3], 3, 2, maxsize, does_not_raise()),
+        ([2, 1, 3], 2, 2, maxsize, pytest.raises(ValueError)),  # noqa: PT011
+        ([2, 1, 3], 1, 2, maxsize, pytest.raises(ValueError)),  # noqa: PT011
+        ([2, 1, 3], 1, 3, maxsize, pytest.raises(ValueError)),  # noqa: PT011
+        ([2, 1, 3], 1, 0, 1, pytest.raises(ValueError)),  # noqa: PT011
+        ([2, 1, 3], 3, 0, 1, pytest.raises(ValueError)),  # noqa: PT011
+    ],
+)
+def test_index(
+    lst: list,
+    value: Any,
+    start: int,
+    stop: int,
+    expectation: AbstractContextManager,
+) -> None:
+    dq = Deque(lst)
+    ans = deque(lst)
+
+    aindex = 0
+    with expectation:
+        aindex = ans.index(value, start, stop)
+
+    dindex = 0
+    with expectation:
+        dindex = dq.index(value, start, stop)
+
+    assert dindex == aindex
+
+
+def test_insert() -> None:
+    dq = Deque(maxlen=None)
+    deq: deque[int] = deque(maxlen=None)
+
+    for idx, val in [
+        (1, 12),
+        (-1, 24),
+        (2, 13),
+        (-2, 67),
+        (10, 123),
+        (-10, 321),
+    ]:
+        dq.insert(idx, val)
+        deq.insert(idx, val)
+        assert dq == deq
+
+
+def test_insert_with_maxlen() -> None:
+    maxlen = 5
+    dq = Deque(maxlen=maxlen)
+    deq: deque[int] = deque(maxlen=maxlen)
+
+    for idx, val in [
+        (1, 12),
+        (-1, 24),
+        (2, 13),
+        (-2, 67),
+        (0, 45),
+    ]:
+        dq.insert(idx, val)
+        deq.insert(idx, val)
+        assert dq == deq
+
+    with pytest.raises(IndexError):
+        deq.insert(0, -12)
+
+    with pytest.raises(IndexError):
+        dq.insert(0, -12)
+
+    maxlen = 0
+    dq = Deque(maxlen=maxlen)
+    deq = deque(maxlen=maxlen)
+
+    with pytest.raises(IndexError):
+        deq.insert(0, -12)
+
+    with pytest.raises(IndexError):
+        dq.insert(0, -12)
+
+
+def test_popleft() -> None:
+    it = deque(random.sample(range(100), 10))
+    dq = Deque(it)
+
+    while dq:
+        dq.popleft()
+        it.popleft()
+        assert dq == it
+
+    with pytest.raises(IndexError):
+        dq.popleft()
+
+    with pytest.raises(IndexError):
+        dq.popleft()
+
+
+def test_pop() -> None:
+    it = deque(random.sample(range(100), 10))
+    dq = Deque(it)
+
+    while dq:
+        dq.pop()
+        it.pop()
+        assert dq == it
+
+    with pytest.raises(IndexError):
+        dq.pop()
+
+    with pytest.raises(IndexError):
+        dq.pop()
+
+
+def test_remove() -> None:
+    it = deque(random.sample(range(100), 10))
+    dq = Deque(it)
+
+    while dq:
+        item = random.choice(it)
+
+        dq.remove(item)
+        it.remove(item)
+
+        assert dq == it
+
+
+@pytest.mark.parametrize("it", [[], [0], [1, 2], [1, -1, 0]])
+def test_reversing(it: Iterable) -> None:
+    icopy = deque(it)
+    dq = Deque(icopy)
+    rev = tuple(reversed(icopy))
+
+    assert tuple(reversed(dq)) == rev
+
+    dq.reverse()
+    assert dq == rev
+
+
+@pytest.mark.parametrize(
+    ("lst", "ns"),
+    [
+        ([], [-1, 0, 1]),
+        ([1], list(range(-2, 3))),
+        ([2, -1, 0, 3], list(range(-5, 6))),
+    ],
+)
+def test_rotate(lst: list, ns: list[int]) -> None:
+    dq = Deque(lst)
+    deq = deque(lst)
+
+    for n in ns:
+        dq.rotate(n)
+        deq.rotate(n)
+
+        assert dq == deq
